updated gateway key, added authentication and music usecases, added navigation_main.xml to MainActivity.kt, finished fetching data

+++ app/src/main/java/com/bsamy/musix/model/dtos/music/StatisticsDto.kt
@@ -9,5 +9,5 @@ data class StatisticsDto(
     @SerializedName("estimatedTotalCount")
     val estimatedTotalCount: Int?,
     @SerializedName("popularity")
-    val popularity: Int?
+    val popularity: Float?
 )
